requirements.txt
asgiref==3.9.2
Django==5.2.6
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
